requirements.txt
aiogram==3.8.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1